fix: improve coding style
