+++ src/shared/components/article-summary/article-summary-form/article-summary-form.component.tsx
@@ -4,13 +4,13 @@ import LinkIcon from '@/shared/icons/link.icon'
 import CustomButton from '../../custom-button/custom-button.component'
 import { CustomButtonTypes } from '../../custom-button/custom-button.types'
 
-const SummaryForm = () => {
+const ArticleSummaryForm = () => {
     const handleInput = () => {}
 
     const handleSubmit = () => {}
 
     return (
-        <form action="" className="relative">
+        <div className="relative">
             <LinkIcon />
             <CustomInput
                 type="url"
@@ -19,11 +19,11 @@ const SummaryForm = () => {
                 changeEvent={handleInput}
                 className="block w-full rounded-md border border-gray-200 bg-white py-2.5 pl-10 pr-12 text-sm shadow-lg font-medium focus:border-black focus:outline-none focus:ring-0 peer"
             />
-            <CustomButton type={CustomButtonTypes.SUBMIT} onClick={handleSubmit} buttonType="submit">
+            <CustomButton type={CustomButtonTypes.SUBMIT} onClick={handleSubmit}>
                 ↵
             </CustomButton>
-        </form>
+        </div>
     )
 }
 
-export default SummaryForm
+export default ArticleSummaryForm
