+++ src/shared/components/article-summary/article-summary.types.ts
@@ -2,7 +2,7 @@ export interface HistoryProps {
     history: Array<string>
 }
 
-export interface SummaryDetailProps {
+export interface ArticleSummaryDetailProps {
     loading?: boolean
     summary?: string
     error?: string
