+++ src/shared/icons/copy.icon.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react'
 import { IconProps } from '../types/icon.types'
 
-const CopyIcon: FC<IconProps> = () => {
+const CopyIcon: FC<IconProps> = ({ width = '40%', height = '40%' }) => {
     return (
         <svg
             fill="none"
@@ -11,8 +11,8 @@ const CopyIcon: FC<IconProps> = () => {
             strokeLinejoin="round"
             strokeWidth="1.5"
             viewBox="0 0 24 24"
-            width="40%"
-            height="40%"
+            width={width}
+            height={height}
             className="object-contain"
             version="1.1"
             id="svg4"
