+++ src/shared/icons/loader.icon.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 
-const LoaderIcon = () => {
+const LoaderIcon = ({ width = 200, height = 200 }) => {
     return (
         <svg
             xmlns="http://www.w3.org/2000/svg"
-            width="200px"
-            height="200px"
+            width={width}
+            height={height}
             viewBox="0 0 100 100"
             preserveAspectRatio="xMidYMid"
         >
